sales order done?

+++ src/app/modules/transactions/pages/sales-order/sales-order-add/sales-order-header/sales-order-header.page.ts
@@ -1,15 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
+import { FormGroup, UntypedFormBuilder, Validators } from '@angular/forms';
 import { ActionSheetController, ModalController, NavController } from '@ionic/angular';
 import { Customer } from 'src/app/modules/transactions/models/customer';
 import { SalesOrderService } from 'src/app/modules/transactions/services/sales-order.service';
 import { AuthService } from 'src/app/services/auth/auth.service';
 import { ToastService } from 'src/app/services/toast/toast.service';
 import { MasterListDetails } from 'src/app/shared/models/master-list-details';
-import { ModuleControl } from 'src/app/shared/models/module-control';
 import { PrecisionList } from 'src/app/shared/models/precision-list';
 import { SearchDropdownList } from 'src/app/shared/models/search-dropdown-list';
-import { SearchDropdownPage } from 'src/app/shared/pages/search-dropdown/search-dropdown.page';
 
 @Component({
   selector: 'app-sales-order-header',
@@ -18,11 +16,12 @@ import { SearchDropdownPage } from 'src/app/shared/pages/search-dropdown/search-
 })
 export class SalesOrderHeaderPage implements OnInit {
 
-  objectForm: UntypedFormGroup;
+  objectForm: FormGroup;
 
-  moduleControl: ModuleControl[] = [];
-
-  transactionShowDetailHeader: boolean = true;
+  customPopoverOptions: any = {
+    message: 'Select one',
+    cssClass: 'popover-in-modal'
+  };
 
   constructor(
     private authService: AuthService,
@@ -36,75 +35,12 @@ export class SalesOrderHeaderPage implements OnInit {
     this.newForm();
   }
 
-  ngOnInit() {
-    this.loadCustomerList();
-    this.loadMasterList();
-    this.loadModuleControl();
-  }
-  
-  customers: Customer[] = [];
-  selectedCustomer: Customer;
-  customerSearchDropdownList: SearchDropdownList[] = [];
-  loadCustomerList() {
-    this.salesOrderService.getCustomerList().subscribe(response => {
-      this.customers = response;
-      this.customers = this.customers.filter(r => r.businessModelType === 'T');
-      this.customers.forEach(r => {
-        this.customerSearchDropdownList.push({
-          id: r.customerId,
-          code: r.customerCode,
-          description: r.name
-        })
-      })
-    }, error => {
-      console.log(error);
-    })
-  }
-
-  customerMasterList: MasterListDetails[] = [];
-  locationMasterList: MasterListDetails[] = [];
-  currencyMasterList: MasterListDetails[] = [];
-  countryMasterList: MasterListDetails[] = [];
-  termPeriodMasterList: MasterListDetails[] = [];
-  salesAgentMasterList: MasterListDetails[] = [];
-  loadMasterList() {
-    this.salesOrderService.getMasterList().subscribe(response => {
-      this.customerMasterList = response.filter(x => x.objectName == 'Customer').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.locationMasterList = response.filter(x => x.objectName == 'Location').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.currencyMasterList = response.filter(x => x.objectName == 'Currency').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.countryMasterList = response.filter(x => x.objectName == 'Country').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.termPeriodMasterList = response.filter(x => x.objectName == 'TermPeriod').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.salesAgentMasterList = response.filter(x => x.objectName == 'SalesAgent').flatMap(src => src.details).filter(y => y.deactivated == 0);
-      this.mapSearchDropdownList();
-    }, error => {
-      console.log(error);
-    })
-  }
-
-  precisionSales: PrecisionList = { precisionId: null, precisionCode: null, description: null, localMin: null, localMax: null, foreignMin: null, foreignMax: null, localFormat: null, foreignFormat: null };
-  precisionTax: PrecisionList = { precisionId: null, precisionCode: null, description: null, localMin: null, localMax: null, foreignMin: null, foreignMax: null, localFormat: null, foreignFormat: null };
-  loadModuleControl() {
-    this.authService.moduleControlConfig$.subscribe(obj => {
-      this.moduleControl = obj;
-      let transactionShowDetailHeader = this.moduleControl.find(r => r.ctrlName === "TransactionShowDetailHeader")?.ctrlValue;
-      if (transactionShowDetailHeader) {
-        this.transactionShowDetailHeader = transactionShowDetailHeader === '1' ? true : false;
-      }
-    }, error => {
-      console.log(error);
-    })
-    this.authService.precisionList$.subscribe(precision =>{
-      this.precisionSales = precision.find(x => x.precisionCode == "SALES");
-      this.precisionTax = precision.find(x => x.precisionCode == "TAX");
-    })
-  }
-
   newForm() {
     this.objectForm = this.formBuilder.group({
       salesOrderId: [0],
       salesOrderNum: [null],
       salesAgentId: [null],
-      trxDate: [null],
+      trxDate: [new Date()],
       typeCode: [null],
       customerId: [null, [Validators.required]],
       shipAddress: [null, [Validators.maxLength(500)]],
@@ -112,8 +48,8 @@ export class SalesOrderHeaderPage implements OnInit {
       shipPhone: [null],
       shipEmail: [null, [Validators.email]],
       shipFax: [null],
-      shipAreaId:[null],
-      shipMethodId:[null],
+      shipAreaId: [null],
+      shipMethodId: [null],
       attention: [null],
       termPeriodId: [null],
       locationId: [null],
@@ -134,7 +70,7 @@ export class SalesOrderHeaderPage implements OnInit {
       masterUDGroup3: [null],
       isItemPriceTaxInclusive: [null],
       isDisplayTaxInclusive: [null],
-      sourceType: [null],
+      sourceType: ['M'],
       businessModelType: [null],
       remark: [null],
       isHomeCurrency: [null],
@@ -143,51 +79,59 @@ export class SalesOrderHeaderPage implements OnInit {
     });
   }
 
-  locationSearchDropdownList: SearchDropdownList[] = [];
-  mapSearchDropdownList() {
-    this.locationMasterList.forEach(r => {
-      this.locationSearchDropdownList.push({
-        id: r.id,
-        code: r.code,
-        description: r.description
-      })
+  ngOnInit() {
+    this.loadCustomerList();
+    this.loadMasterList();
+    this.loadModuleControl();
+  }
+  
+  precisionSales: PrecisionList = { precisionId: null, precisionCode: null, description: null, localMin: null, localMax: null, foreignMin: null, foreignMax: null, localFormat: null, foreignFormat: null };
+  precisionTax: PrecisionList = { precisionId: null, precisionCode: null, description: null, localMin: null, localMax: null, foreignMin: null, foreignMax: null, localFormat: null, foreignFormat: null };
+  loadModuleControl() {    
+    this.authService.precisionList$.subscribe(precision =>{
+      this.precisionSales = precision.find(x => x.precisionCode == "SALES");
+      this.precisionTax = precision.find(x => x.precisionCode == "TAX");
     })
   }
 
-  async showLocationSearchDropdown() {
-    let dropdownlist: SearchDropdownList[] = [];
-    // change source
-    this.locationMasterList.forEach(r => {
-      dropdownlist.push({
-        id: r.id,
-        code: r.code,
-        description: r.description
-      })
+  customerMasterList: MasterListDetails[] = [];
+  locationMasterList: MasterListDetails[] = [];
+  currencyMasterList: MasterListDetails[] = [];
+  loadMasterList() {
+    this.salesOrderService.getMasterList().subscribe(response => {
+      this.customerMasterList = response.filter(x => x.objectName == 'Customer').flatMap(src => src.details).filter(y => y.deactivated == 0);
+      this.locationMasterList = response.filter(x => x.objectName == 'Location').flatMap(src => src.details).filter(y => y.deactivated == 0);
+      this.currencyMasterList = response.filter(x => x.objectName == 'Currency').flatMap(src => src.details).filter(y => y.deactivated == 0);
+    }, error => {
+      console.log(error);
     })
-    const modal = await this.modalController.create({
-      component: SearchDropdownPage,
-      componentProps: {
-        searchDropdownList: dropdownlist
-      },
-      canDismiss: true
-    });
-    await modal.present();
-    let { data } = await modal.onWillDismiss();
-    if (data) {
-      this.objectForm.patchValue({ locationId: data.id });
-    }
   }
 
-  locationChanged(event) {
-    this.objectForm.patchValue({ locationId: event.id });
+  customers: Customer[] = [];
+  selectedCustomer: Customer;
+  customerSearchDropdownList: SearchDropdownList[] = [];
+  loadCustomerList() {
+    this.salesOrderService.getCustomerList().subscribe(response => {
+      this.customers = response;
+      this.customers = this.customers.filter(r => r.businessModelType === 'T');
+      this.customers.forEach(r => {
+        this.customerSearchDropdownList.push({
+          id: r.customerId,
+          code: r.customerCode,
+          description: r.name
+        })
+      })
+    }, error => {
+      console.log(error);
+    })
   }
 
   selectedCustomerLocationList: MasterListDetails[] = [];
   onCustomerSelected(event) {
-    if (event) {
-      this.salesOrderService.removeItems();
-      var lookupValue = this.customerMasterList?.find(e => e.id == event.id);
+    if (event && event !== undefined) {
+      var lookupValue = this.customerMasterList?.find(e => e.id === event.id);
       if (lookupValue != undefined) {
+        this.salesOrderService.removeItems();
         this.objectForm.patchValue({ customerId: lookupValue.id });
         this.objectForm.patchValue({ businessModelType: lookupValue.attribute5 });
         if (lookupValue.attributeArray1.length > 0) {
@@ -195,17 +139,17 @@ export class SalesOrderHeaderPage implements OnInit {
         } else {
           this.selectedCustomerLocationList = [];
         }
-        this.objectForm.patchValue({
-          salesAgentId: parseFloat(lookupValue.attribute1),
-          termPeriodId: parseFloat(lookupValue.attribute2),
-          countryId: parseFloat(lookupValue.attribute3),
-          currencyId: parseFloat(lookupValue.attribute4),
-          locationId: parseFloat(lookupValue.attribute6),
-          toLocationId: null,
-          isItemPriceTaxInclusive: lookupValue.attribute8 == '1' ? true : false,
-          isDisplayTaxInclusive: lookupValue.attribute9 == '1' ? true : false
+        this.objectForm.patchValue({ 
+          salesAgentId: parseFloat(lookupValue.attribute1), 
+          termPeriodId: parseFloat(lookupValue.attribute2), 
+          countryId: parseFloat(lookupValue.attribute3), 
+          currencyId: parseFloat(lookupValue.attribute4), 
+          locationId: parseFloat(lookupValue.attribute6), 
+          toLocationId: null, 
+          isItemPriceTaxInclusive: lookupValue.attribute8 == '1' ? true : false, 
+          isDisplayTaxInclusive: lookupValue.attribute9 == '1' ? true : false 
         });
-        this.onCurrencySelected({ id: lookupValue.attribute4 });
+        this.onCurrencySelected(lookupValue.attribute4);
         if (lookupValue.attribute5 == "T") {
           this.objectForm.controls.toLocationId.clearValidators();
           this.objectForm.controls.toLocationId.updateValueAndValidity();
@@ -225,7 +169,7 @@ export class SalesOrderHeaderPage implements OnInit {
 
   onCurrencySelected(event: any) {
     if (event) {
-      var lookupValue = this.currencyMasterList?.find(e => e.id == event.id);
+      var lookupValue = this.currencyMasterList?.find(e => e.id == event);
       if (lookupValue != undefined) {
         this.objectForm.patchValue({ currencyRate: parseFloat(lookupValue.attribute1) });
         if (lookupValue.attribute2 == "Y") {
@@ -256,24 +200,16 @@ export class SalesOrderHeaderPage implements OnInit {
         }]
     });
     await actionSheet.present();
-
     const { role } = await actionSheet.onWillDismiss();
-
     if (role === 'confirm') {
       this.salesOrderService.resetVariables();
       this.navController.navigateBack('/transactions/sales-order');
     }
   }
 
-  nextStep() {
-    this.objectForm.getRawValue
-    console.log("🚀 ~ file: sales-order-header.page.ts ~ line 272 ~ SalesOrderHeaderPage ~ nextStep ~ this.objectForm.value", this.objectForm.value)
-    if (this.objectForm.valid) {
-      this.salesOrderService.setChoosenCustomer(this.objectForm.value);
-      this.navController.navigateForward('/transactions/sales-order/sales-order-item');
-    } else {
-      this.toastService.presentToast('Error', 'Please select customer to continue', 'bottom', 'danger', 1000);
-    }
+  async nextStep() {
+    await this.salesOrderService.setHeader(this.objectForm.value);
+    this.navController.navigateForward('/transactions/sales-order/sales-order-item');    
   }
 
 }
